requirements.txt
aiogram>=3.0
selenium
webdriver-manager
undetected-chromedriver
Pillow
python-dotenv
